Moved MarkLogicConnectorException back to the internal package
